Fix ShadowAccessibilityManager to keep static instance
This is a squashed commit of two commits from changwan@:
 - 6d9859431b1201ad5d4139d0b397dd5b4bb22863
 - b02c65cc6d7465f58f0de48a39914aa905692afa

1) Shadows.shadowOf(mAccessibilityManager).setEnabled(true)
2) AccessibilityManager.getInstance(mContext).isEnabled()

Calling 1) does not make 2) return true, and 2) is the common pattern
that can be found in the Android framework, so we cannot enable
accessibility in robolectric tests correctly.

The reason is that the current implementation creates new instance
each time, while setEnabled() and other methods only apply to
the current instance. This can be fixed by keeping a static instance, an
approach similar to AccessibilityManager.getInstance().

BUG=756707